+++ folenski/staff/Models/Change.php
@@ -5,6 +5,7 @@
  *
  * @author  folenski
  * @version 1.0 17/12/2022: Version Initiale 
+ * @version 1.1 09/07/2023: propriété _error ajoutée pour supprimer un warning
  * 
  */
 
@@ -22,6 +23,7 @@ final class Change implements TableInterface
         "created_at" => "DATETIME NOT NULL",
         "updated_at" => "DATETIME NOT NULL",
     ];
+    private string|bool $_error = false;
 
     /**
      * @return array retourne le nom et la description de la table
@@ -47,7 +49,7 @@ final class Change implements TableInterface
      */
     function errors(): false|string
     {
-        return (isset($this->_error)) ? $this->_error : false;
+        return $this->_error;
     }
 
     /**