+++ folenski/staff/Models/Environment.php
@@ -7,6 +7,8 @@
  * @version 1.0  4/08/2022: version Initiale 
  * @version 1.1 10/12/2022: utilisation des tags
  * @version 1.2 26/12/2022: utilisation des types small pour mysql
+ * @version 1.3 10/01/2023: suppression de la colonne j_contact
+ * @version 1.4 09/07/2023: propriété _error ajoutée pour supprimer un warning
  * 
  */
 
@@ -21,12 +23,12 @@ final class Environment implements TableInterface
     private const _DESC = [
         "name"       => "VARCHAR(3) PRIMARY KEY",
         "j_option"   => "#JSON_SM NOT NULL",
-        "j_contact"  => "#JSON_SM",
         "j_index"    => "#JSON_SM NOT NULL",
         "j_route"    => "#JSON_SM NOT NULL",
         "created_at" => "DATETIME NOT NULL",
         "updated_at" => "DATETIME NOT NULL",
     ];
+    private string|bool $_error = false;
 
     /**
      * @return array retourne le nom et la description de la table
@@ -43,11 +45,10 @@ final class Environment implements TableInterface
      */
     function check(array $fields): array|false
     {
-        [$ret, $fail, $name, $j_option, $j_contact, $j_index, $j_route]
+        [$ret, $fail, $name, $j_option, $j_index, $j_route]
             = Carray::arrayCheck($fields, [
                 "name" => ["type" => "string"],
                 "j_option" => ["json" => true],
-                "j_contact" => ["json" => true],
                 "j_index" => ["json" => true],
                 "j_route" => ["json" => true]
             ]);
@@ -55,8 +56,8 @@ final class Environment implements TableInterface
             $this->_error = $fail;
             return false;
         }
-        unset($this->_error);
-        return compact("name", "j_option", "j_contact", "j_index", "j_route");
+        $this->_error = false;
+        return compact("name", "j_option", "j_index", "j_route");
     }
 
     /**
@@ -64,7 +65,7 @@ final class Environment implements TableInterface
      */
     function errors(): false|string
     {
-        return (isset($this->_error)) ? $this->_error : false;
+        return $this->_error;
     }
 
     /**